pyproject.toml
[project]
name = "mappy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "tensorboard>=2.14.0",
    "tqdm>=4.66.0",
    "flask>=3.0.0",
    "pyyaml>=6.0",
]
